Finishing V1
went with a different design language with this one as I did a hacker man inspiration with this as it fits well this the logo, make the h1 bigger and have the logos positions also perfect with the absolute positions. Put the social logos outside the main "Row" div to not have it influenced by the "Row" settings.